Decimal quantities and beautified Edit meal item page.

+++ Data/MealItemIngredient.cs
@@ -25,7 +25,8 @@ namespace clean_aspnet_mvc.Data
         public MealItem MealItem { get; set; }
 
         [Required]
-        public int Quantity { get; set; }
+
+        public decimal Quantity { get; set; }
         [Display(Name = "Measure Type")]
         public string MeasureType { get; set; }
